Reordered pages, starts on grass field, then loops around after living reduction
