+++ Components/Header.tsx
@@ -15,7 +15,7 @@ const Header = () => {
       </Head>
 
       <div className={styles.title_container}>
-        <Link href="/" className={styles.homeLink}><img width="75" height="75" src="/titleicon.svg"></img></Link>
+        <Link href="/particle_image" className={styles.homeLink}><img width="75" height="75" src="/titleicon.svg"></img></Link>
         <div className={styles.logo_block}>
           <a href="https://github.com/chloenott"><img width="35" height="35" src="/icons/iconmonstr-github-2.svg" alt="GitHub logo"></img></a>
           <a href="https://www.linkedin.com/in/chloenott/"><img width="35" height="35" src="/icons/iconmonstr-linkedin-2.svg" alt="LinkedIn logo"></img></a>
